requirements.txt
fastapi
uvicorn
requests
torch
python-dotenv
pytest
numpy